pyproject.toml
[project]
name = "aa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper==20250625",
    "pipwin==0.5.2",
    "pocketsphinx==5.0.4",
    "pyaudio==0.2.14",
    "pyttsx3==2.99",
    "soundfile==0.13.1",
    "speechrecognition==3.14.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "googlesearch-python>=1.3.0",
    "helium>=5.0.0",
    "pillow>=10.3.0",
    "pydantic-ai>=0.7.3",
    "selenium>=4.22.0",
    "webdriver-manager>=4.0.2",
    "psutil>=7.1.0",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.6.0",
    "scrapy>=2.13.3",
    "scrapy-playwright>=0.0.44",
    "pyautogui>=0.9.54",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    "pytz>=2025.2",
]

[tool.uv.workspace]
members = ["mcp_agent"]
